// src/modules/user/generated-types/user-resolvers-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
};

export enum Color {
  Blue = 'BLUE',
  Red = 'RED'
}

export type LoginInput = {
  email: Scalars['String'];
  password: Scalars['String'];
};

export type LoginResponse = {
  __typename?: 'LoginResponse';
  expiresAt: Scalars['Date'];
  id: Scalars['Int'];
  roles?: Maybe<Array<Scalars['String']>>;
  token: Scalars['String'];
};
// ...
export type RegisterInput = {
  email: Scalars['String'];
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  password: Scalars['String'];
  roleName: Scalars['String'];
};